backend/initMinIO/requirements.txt
minio==7.2.15
python-dotenv==1.1.1